Facebook synclet tests vows -> mocha.
Ported over the Facebook vows tests to mocha, bringing over the fixtures
as well. This involved fixing at least one serious bug in the Facebook
photos synclet, so yay for that.

Two-phase synclets that do entirely different things on different passes
are very confusing and should probably be made less so.