Remove rando test file, replace with test cases.
This is the start at an effort to test the OAuth provider for 3rd party
clients. It takes the sample express app @quartzjer and @smurthas
created and crudely shoves it into a Mocha wrapper, and then tries to
test against that.

Unfortunately, the way this all works is that the sample app proxies the
auth request to the locker, which in turn then goes and *actually
authorizes against Twitter or Facebook and gets back an (eventually
per-user) access token*. @kristjan and I discussed how to mock all this
out, but since the test case is starting up a self-contained Express app
*and* an instance of the locker (all running in the same process space,
no less), the two, equally unappetizing options were either to use
Fakeweb to mock out all the requests, thus turning the whole thing into
a hall of mirrors, or to integrate something like Sinon and inject in
some selective mocking, which would be hairy and gross.

Instead, I'm committing this commented code (and nuking the sample test
app, cuz come on, that's not going to integrate with CI) and maybe going
to write some test cases against the actual OAuth provider code itself.