Introducing DAL and IJOD on S3
DAL is intended to become a database abstraction layer that will allow
us to have a consistent API for multiple backends.  Right now it's just
mysql for our RDS use, but sqlite is intended as an alternative to give
a more single user friendly option.

IJOD now uses DAL to talk to RDS and do all of it's DB operations, and
saves the files to S3 as a time based series.

This is not fully working, in fact retrieving records is disabled, but
allows others to see and understand progress.