Give accurate feedback when deleting profiles
Fixes #409

We weren't noticing whether the profile was a valid one to actually delete, and
returning `true` in any non-error case. Now, you'll get:

* Missing/bad access token: 401
* Account doesn't exist: 404
* No delete parameter: 400
* Account wasn't deleted: false
* New - Profile doesn't exist: 404
* New - Profile wasn't deleted: false
* Account or profile was really deleted: true

The mix of !200/message with 200/boolean is weird, but at least one person's
checking the boolean so I don't want to change it in this version. It would be
more consistent to let HTTP 200 mean "true" and always have a message by
returning the ID of the entity that was just deleted.