Create posting endpoint with basic tests
I put the posting code into its own module so I can load it independantly of a
running server. Also playing around with a way to easily mock the request and
verify the response.

The mock request simply wraps your data object so the server can use Express's
`params` function.

The mock response contains the functions we expect the server to call and uses
the to verify what they're called with.